test: remove unnecessary changes
